moved delete recipe files to right file name
