+++ src/backend/service/delete_recipe/DeleteRecipeController.java
@@ -1,4 +1,4 @@
-package backend.service.delete_cookbook;
+package backend.service.delete_recipe;
 
 public class DeleteRecipeController {
