+++ src/backend/service/delete_recipe/DeleteRecipeDAI.java
@@ -1,4 +1,4 @@
-package backend.service.delete_cookbook;
+package backend.service.delete_recipe;
 
 import backend.entity.Recipe;
