+++ src/backend/service/delete_recipe/DeleteRecipeDeleteDAI.java
@@ -1,4 +1,4 @@
-package backend.service.delete_cookbook;
+package backend.service.delete_recipe;
 
 import backend.entity.Cookbook;
 import backend.entity.Recipe;
