+++ src/backend/service/delete_recipe/DeleteRecipeInputBoundary.java
@@ -1,4 +1,4 @@
-package backend.service.delete_cookbook;
+package backend.service.delete_recipe;
 
 public interface DeleteRecipeInputBoundary {
     void execute(DeleteRecipeInputData deleteRecipeInputData);
