+++ src/backend/service/delete_recipe/DeleteRecipeInteractor.java
@@ -1,4 +1,4 @@
-package backend.service.delete_cookbook;
+package backend.service.delete_recipe;
 
 public class DeleteRecipeInteractor implements DeleteRecipeInputBoundary {
