+++ src/backend/service/delete_recipe/DeleteRecipeOutputBoundary.java
@@ -1,4 +1,4 @@
-package backend.service.delete_cookbook;
+package backend.service.delete_recipe;
 
 public interface DeleteRecipeOutputBoundary {
     void prepareSuccessView(DeleteRecipeOutputData deleteRecipeOutputData);
