+++ src/backend/service/delete_recipe/DeleteRecipeOutputData.java
@@ -1,4 +1,4 @@
-package backend.service.delete_cookbook;
+package backend.service.delete_recipe;
 
 public class DeleteRecipeOutputData {
     private final String recipeName;
