+++ src/backend/service/delete_recipe/DeleteRecipePresenter.java
@@ -1,4 +1,4 @@
-package backend.service.delete_cookbook;
+package backend.service.delete_recipe;
 
 public class DeleteRecipePresenter implements DeleteRecipeOutputBoundary {
     @Override
